requirements.txt
google-generativeai
fastapi
uvicorn
streamlit
openai
pypdf
python-multipart
python-dotenv
requests
pydantic
